回学校了，我的第26次java学习,函数式接口有什么用啊？多此一举。。。。。

+++ Block_Horse/src/方法/overLoad.java
@@ -36,7 +36,7 @@ public class overLoad {
         return a + b + c;
     }
 
-    public static int sum(int a, int b, int c, int d) {
+    public static long sum(int a, int b, int c, long d) {
         return a + b + c + d;
     }
